fix(node): ensure aggregate blob info is not deleted prematurely
There was a bug where in a particular corner case, an aggregate blob
info was deleted while some per-object blob info entries still existed
(see details below). If then the garbage collection attempted to update
the aggregate blob info for the expired object, it caused a panic in the
merge operator.

This fixes the bug by adding a check to delete the aggregate blob info
only if no corresponding per-object blob info entries still exist.

The bug was triggered in the `simtest_test_lagging_node_recovery` test
as follows:
- A blob was registered twice as deletable in epoch 2 with an end epoch
  of 4.
- One of the corresponding objects was deleted in epoch 3.
- The deletion event was processed at the beginning of epoch 4. At this
  point, the blob was already expired, meaning that blob data could be
  deleted.
- However, the second deletable object was still stored in the
  per-object blob info table as the clean-up was not yet complete.

Besides fixing the bug, this commit also adds a consistency check for
invariants of the blob info tables (turned off by default in the config)
and improves some logging.

+++ crates/walrus-service/src/node/storage.rs
@@ -668,7 +668,7 @@ impl Storage {
                 }
             };
 
-            if blob_info.is_registered(current_epoch) {
+            if blob_info.can_data_be_deleted(current_epoch) {
                 tracing::trace!(
                     %blob_id,
                     "skipping blob that is still registered",
@@ -757,7 +757,7 @@ impl Storage {
             );
             return Ok(false);
         };
-        if blob_info.is_registered(current_epoch) {
+        if blob_info.can_data_be_deleted(current_epoch) {
             tracing::debug!(
                 %blob_id,
                 "blob was reregistered before attempting to delete expired blob data",
@@ -766,10 +766,14 @@ impl Storage {
         }
 
         // At this point we are sure that the blob is no longer registered and can actually delete
-        // the blob info and data. If the blob is reregistered outside this transaction, the
-        // transaction will fail.
-        self.blob_info
-            .delete_in_transaction(&transaction, blob_id)?;
+        // the data. If the blob is reregistered outside this transaction, the transaction will
+        // fail.
+        if blob_info.can_blob_info_be_deleted(current_epoch) {
+            // It is possible that there are still (expired) deletable blob objects for this blob
+            // ID. In that case, we should not delete the aggregate blob info yet.
+            self.blob_info
+                .delete_in_transaction(&transaction, blob_id)?;
+        }
         self.delete_blob_data_in_transaction(&transaction, blob_id, shards)?;
 
         if let Err(error) = transaction.commit() {
@@ -1044,6 +1048,13 @@ impl Storage {
     ) -> PerObjectBlobInfoIterator<'_> {
         self.blob_info
             .certified_per_object_blob_info_iter_before_epoch(epoch, std::ops::Bound::Unbounded)
+    }
+
+    /// Checks internal invariants of the blob info table.
+    pub(crate) fn blob_info_invariants_check(&self) {
+        if let Err(error) = self.blob_info.check_invariants() {
+            tracing::error!(?error, "blob info internal consistency check failed");
+        }
     }
 
     /// Returns the current event cursor.
